Replace class list check with data attribute
For mouse events, when clicking on buttons with icons, the class list
for e.target can be either the icon class list or button class list.

e.currentTarget can be used to ensure the listener fires for the button
even when the icon is pressed. However, to simplify things, replace this
with a data attribute check instead.